+++ crates/sncast/src/helpers/mod.rs
@@ -1,3 +1,3 @@
-pub mod config;
+pub mod configuration;
 pub mod constants;
 pub mod scarb_utils;
